Refactoring future baseline tests
